Factor out getInterpreterTypeName() and sortInterpreters().
